voice_to_notes_generator/requirements.txt
python-dotenv
google-generativeai
groq
python-docx